Move extensions to their own folder.
Make the class partial an create cs files for a group of extensions
